requirements.txt
hikari[speedups]
hikari-lightbulb
hikari-miru
python-dotenv
uvloop
